requirements.txt
antspyx==0.2.4
boltons==20.0.0
gpustat==0.6.0
imutils==0.5.3
Keras==2.4.3
Keras-Preprocessing==1.1.0
matplotlib==3.2.1
nibabel==3.1.1
numpy==1.18.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python==4.4.0.40
opencv-python-headless==4.2.0.32
pandas==1.1.0
protobuf==3.11.3
psutil==5.7.0
redlock-py==1.0.8
scikit-learn==0.23.2
tensorflow-gpu==2.3.1
toml==0.10.0
tornado==6.0.4
tqdm==4.44.1
wandb==0.9.5
